+++ flake8_scrapy/_finders/validation.py
@@ -1,12 +1,8 @@
-from __future__ import annotations
-
 import ast
 import json
 from pathlib import Path
 
-from packaging.requirements import InvalidRequirement, Requirement
-from packaging.utils import canonicalize_name
-from packaging.version import InvalidVersion, Version
+from .versions import is_version_greater_than
 
 
 def is_valid_log_level(value) -> bool:
@@ -46,205 +42,7 @@ def looks_like_class_import_path(value: str) -> bool:
     return parts[-1][0].isupper()
 
 
-def looks_like_callable_import_path(value: str) -> bool:
-    """Check if a string looks like a valid import path for any callable (function, class, etc.)."""
-    if not value:
-        return False
-    parts = value.split(".")
-    MINIMUM_IMPORT_PARTS = 2
-    if len(parts) < MINIMUM_IMPORT_PARTS:
-        return False
-    return all(part.isidentifier() for part in parts)
-
-
-def parse_requirement_line(line: str) -> Requirement | None:
-    """Parse a requirement line and return a Requirement object if valid."""
-    line = line.strip()
-    if not line or line.startswith("#"):
-        return None
-    if line.startswith("-"):
-        return None
-    if "#" in line:
-        line = line.split("#")[0].strip()
-    try:
-        return Requirement(line)
-    except InvalidRequirement:
-        # This handles URLs, local paths, and other non-standard formats
-        return None
-
-
-def is_frozen_requirement(req: Requirement) -> bool:
-    """Check if a requirement is frozen (has an exact version pin)."""
-    return len(req.specifier) == 1 and next(iter(req.specifier)).operator == "=="
-
-
-def is_version_less_than(version: str | Version, min_version: str | Version) -> bool:
-    """Compare two versions, returning True if version < min_version."""
-    try:
-        v1 = version if isinstance(version, Version) else Version(version)
-        v2 = min_version if isinstance(min_version, Version) else Version(min_version)
-        return v1 < v2
-    except InvalidVersion:
-        return False
-
-
-def is_version_greater_than(version: str | Version, min_version: str | Version) -> bool:
-    """Compare two versions, returning True if version > min_version."""
-    try:
-        v1 = version if isinstance(version, Version) else Version(version)
-        v2 = min_version if isinstance(min_version, Version) else Version(min_version)
-        return v1 > v2
-    except InvalidVersion:
-        return False
-
-
-def is_version_less_than_or_equal(
-    version: str | Version, max_version: str | Version
-) -> bool:
-    """Compare two versions, returning True if version <= max_version."""
-    try:
-        v1 = version if isinstance(version, Version) else Version(version)
-        v2 = max_version if isinstance(max_version, Version) else Version(max_version)
-        return v1 <= v2
-    except InvalidVersion:
-        return False
-
-
-def get_package_version_from_requirements(
-    package_name: str, project_root: Path | None
-) -> Version | None:
-    """Get the version of a package from requirements.txt."""
-    if not project_root:
-        return None
-
-    requirements_txt = project_root / "requirements.txt"
-    if not requirements_txt.exists():
-        return None
-
-    try:
-        requirements = requirements_txt.read_text(encoding="utf-8")
-    except (OSError, UnicodeDecodeError):
-        return None
-
-    for line in requirements.splitlines():
-        requirement = parse_requirement_line(line)
-        if requirement is None:
-            continue
-        if not is_frozen_requirement(requirement):
-            continue
-        if canonicalize_name(requirement.name) == canonicalize_name(package_name):
-            version = next(iter(requirement.specifier)).version
-            return Version(version)
-
-    return None
-
-
-def build_package_versions_dict(project_root: Path | None) -> dict[str, Version]:
-    """Build a dictionary of package names to versions from requirements.txt."""
-    package_versions = {}
-    if not project_root:
-        return package_versions
-
-    requirements_txt = project_root / "requirements.txt"
-    if not requirements_txt.exists():
-        return package_versions
-
-    try:
-        requirements = requirements_txt.read_text(encoding="utf-8")
-    except (OSError, UnicodeDecodeError):
-        return package_versions
-
-    for line in requirements.splitlines():
-        requirement = parse_requirement_line(line)
-        if requirement is None:
-            continue
-        if not is_frozen_requirement(requirement):
-            continue
-        version = next(iter(requirement.specifier)).version
-        package_versions[canonicalize_name(requirement.name)] = Version(version)
-
-    return package_versions
-
-
-def format_issue_message(msg_code: str, msg_info: str, detail: str = "") -> str:
-    """Format a standard issue message with optional detail."""
-    base_message = f"{msg_code} {msg_info}"
-    if detail:
-        return f"{base_message}: {detail}"
-    return base_message
-
-
-def format_version_issue_message(
-    msg_code: str, msg_info: str, actual_version: str, minimum_version: str
-) -> str:
-    """Format a version-related issue message."""
-    detail = f"{actual_version} (minimum required: {minimum_version})"
-    return format_issue_message(msg_code, msg_info, detail)
-
-
-def format_replacement_message(
-    msg_code: str, msg_info: str, package_name: str, replacements: list[str]
-) -> str:
-    """Format a package replacement message."""
-    if len(replacements) == 1:
-        replacement_text = replacements[0]
-    else:
-        replacement_text = " or ".join(replacements)
-    detail = f"{package_name} (use {replacement_text} instead)"
-    return format_issue_message(msg_code, msg_info, detail)
-
-
-def check_package_obsolescence(
-    package_name: str, obsolete_packages: dict[str, list[str]]
-) -> tuple[bool, list[str]]:
-    """Check if a package is obsolete and return replacement suggestions."""
-    package_lower = package_name.lower()
-    if package_lower in obsolete_packages:
-        return True, obsolete_packages[package_lower]
-    return False, []
-
-
-def extract_version_from_requirement(req: Requirement) -> str | None:
-    """Extract version string from a frozen requirement."""
-    if not is_frozen_requirement(req):
-        return None
-    return next(iter(req.specifier)).version
-
-
-def validate_feeds_config(  # noqa: PLR0911
-    value_node: ast.AST, feeds_key_versions: dict[str, str], get_package_version_func
-) -> str:
-    """Get specific validation error for FEEDS setting, or empty string if valid."""
-    # FEEDS must be a dict
-    if isinstance(value_node, ast.Constant):
-        value = value_node.value
-        if isinstance(value, dict):
-            return _validate_feeds_dict(
-                value, feeds_key_versions, get_package_version_func
-            )
-        if isinstance(value, str):
-            try:
-                parsed_value = json.loads(value)
-                if not isinstance(parsed_value, dict):
-                    return "must be a dict"
-                return _validate_feeds_dict(
-                    parsed_value, feeds_key_versions, get_package_version_func
-                )
-            except (json.JSONDecodeError, TypeError):
-                return "must be a dict"
-        else:
-            return "must be a dict"
-
-    if isinstance(value_node, ast.Dict):
-        return _validate_feeds_dict_ast(
-            value_node, feeds_key_versions, get_package_version_func
-        )
-
-    # Any other AST node type is invalid for FEEDS
-    return "must be a dict"
-
-
-def _validate_feeds_dict(
+def validate_feeds_dict(
     feeds_dict: dict, feeds_key_versions: dict[str, str], get_package_version_func
 ) -> str:
     """Get validation error for a FEEDS dict value at runtime."""
@@ -258,7 +56,7 @@ def _validate_feeds_dict(
             return f"feed config for {key!r} must be a dict"
 
         # Validate feed config keys and values
-        error = _validate_feed_config(
+        error = validate_feed_config(
             key, feed_config, feeds_key_versions, get_package_version_func
         )
         if error:
@@ -267,58 +65,7 @@ def _validate_feeds_dict(
     return ""
 
 
-def _validate_feeds_dict_ast(
-    dict_node: ast.Dict, feeds_key_versions: dict[str, str], get_package_version_func
-) -> str:
-    """Get validation error for a FEEDS dict AST node."""
-    for key_node, value_node in zip(dict_node.keys, dict_node.values):
-        # Root keys may be strings or Path objects
-        key_repr = "<?>"
-        if isinstance(key_node, ast.Constant):
-            key_repr = repr(key_node.value)
-            if not isinstance(key_node.value, str):
-                return f"key {key_repr} must be a string or Path object"
-        elif not (
-            isinstance(key_node, ast.Call)
-            and isinstance(key_node.func, ast.Name)
-            and key_node.func.id == "Path"
-        ):
-            # Not a string constant or Path() call
-            return f"key {key_repr} must be a string or Path object"
-
-        # Feed config must be a dict
-        if not isinstance(value_node, ast.Dict):
-            # Check if this looks like feed config keys were used at the top level
-            if isinstance(key_node, ast.Constant) and isinstance(key_node.value, str):
-                feed_config_keys = {
-                    "format",
-                    "batch_item_count",
-                    "encoding",
-                    "fields",
-                    "item_classes",
-                    "item_filter",
-                    "indent",
-                    "item_export_kwargs",
-                    "overwrite",
-                    "store_empty",
-                    "uri_params",
-                    "postprocessing",
-                }
-                if key_node.value in feed_config_keys:
-                    return f"missing feed URL: {key_repr} appears to be a feed configuration key, but FEEDS must be a dict where keys are feed URLs (like 'output.json') and values are feed configurations"
-            return f"feed config for {key_repr} must be a dict"
-
-        # Validate feed config AST
-        error = _validate_feed_config_ast(
-            key_repr, value_node, feeds_key_versions, get_package_version_func
-        )
-        if error:
-            return error
-
-    return ""
-
-
-def _validate_feed_config(  # noqa: PLR0911,PLR0912
+def validate_feed_config(  # noqa: PLR0911,PLR0912
     feed_key: str,
     feed_config: dict,
     feeds_key_versions: dict[str, str],
@@ -393,7 +140,91 @@ def _validate_feed_config(  # noqa: PLR0911,PLR0912
     return ""
 
 
-def _validate_feed_config_ast(  # noqa: PLR0911,PLR0912
+def validate_feeds_config(  # noqa: PLR0911
+    value_node: ast.AST, feeds_key_versions: dict[str, str], get_package_version_func
+) -> str:
+    """Get specific validation error for FEEDS setting, or empty string if valid."""
+    # FEEDS must be a dict
+    if isinstance(value_node, ast.Constant):
+        value = value_node.value
+        if isinstance(value, dict):
+            return validate_feeds_dict(
+                value, feeds_key_versions, get_package_version_func
+            )
+        if isinstance(value, str):
+            try:
+                parsed_value = json.loads(value)
+                if not isinstance(parsed_value, dict):
+                    return "must be a dict"
+                return validate_feeds_dict(
+                    parsed_value, feeds_key_versions, get_package_version_func
+                )
+            except (json.JSONDecodeError, TypeError):
+                return "must be a dict"
+        else:
+            return "must be a dict"
+
+    if isinstance(value_node, ast.Dict):
+        return validate_feeds_dict_ast(
+            value_node, feeds_key_versions, get_package_version_func
+        )
+
+    # Any other AST node type is invalid for FEEDS
+    return "must be a dict"
+
+
+def validate_feeds_dict_ast(
+    dict_node: ast.Dict, feeds_key_versions: dict[str, str], get_package_version_func
+) -> str:
+    """Get validation error for a FEEDS dict AST node."""
+    for key_node, value_node in zip(dict_node.keys, dict_node.values):
+        # Root keys may be strings or Path objects
+        key_repr = "<?>"
+        if isinstance(key_node, ast.Constant):
+            key_repr = repr(key_node.value)
+            if not isinstance(key_node.value, str):
+                return f"key {key_repr} must be a string or Path object"
+        elif not (
+            isinstance(key_node, ast.Call)
+            and isinstance(key_node.func, ast.Name)
+            and key_node.func.id == "Path"
+        ):
+            # Not a string constant or Path() call
+            return f"key {key_repr} must be a string or Path object"
+
+        # Feed config must be a dict
+        if not isinstance(value_node, ast.Dict):
+            # Check if this looks like feed config keys were used at the top level
+            if isinstance(key_node, ast.Constant) and isinstance(key_node.value, str):
+                feed_config_keys = {
+                    "format",
+                    "batch_item_count",
+                    "encoding",
+                    "fields",
+                    "item_classes",
+                    "item_filter",
+                    "indent",
+                    "item_export_kwargs",
+                    "overwrite",
+                    "store_empty",
+                    "uri_params",
+                    "postprocessing",
+                }
+                if key_node.value in feed_config_keys:
+                    return f"missing feed URL: {key_repr} appears to be a feed configuration key, but FEEDS must be a dict where keys are feed URLs (like 'output.json') and values are feed configurations"
+            return f"feed config for {key_repr} must be a dict"
+
+        # Validate feed config AST
+        error = validate_feed_config_ast(
+            key_repr, value_node, feeds_key_versions, get_package_version_func
+        )
+        if error:
+            return error
+
+    return ""
+
+
+def validate_feed_config_ast(  # noqa: PLR0911,PLR0912
     feed_key: str,
     dict_node: ast.Dict,
     feeds_key_versions: dict[str, str],
@@ -697,11 +528,12 @@ def _validate_slot_config_ast(slot_key: str, dict_node: ast.Dict) -> str:
     return ""
 
 
-def get_enum_validation_error(
-    setting_name: str, enum_settings: dict[str, list[str]]
-) -> str:
-    """Get enum validation error message for a setting."""
-    if setting_name in enum_settings:
-        allowed_values = ", ".join(f"'{v}'" for v in enum_settings[setting_name])
-        return f"only supports the following values: {allowed_values}."
-    return "only supports specific string values."
+def looks_like_callable_import_path(value: str) -> bool:
+    """Check if a string looks like a valid import path for any callable (function, class, etc.)."""
+    if not value:
+        return False
+    parts = value.split(".")
+    MINIMUM_IMPORT_PARTS = 2
+    if len(parts) < MINIMUM_IMPORT_PARTS:
+        return False
+    return all(part.isidentifier() for part in parts)
